Add fix for data pipeline
As of now, the data waits for ALL collections to refresh the datapoints
for the last 7 days. This can take anywhere from 15 mins up to an hour
(we have ≈1800 collections)

This fix prioritises the dashboard rankings which are the first point of
user interaction. As we are bottlenecked by the API's rate limit, and
the data is only refreshed at 0800 SGT daily, the price points retrieved
at 0800 often have little to no data.

The Collections rankings on the other hand are already indexed by the
various API providers and are more up to date.